refactor: Extrair esquema de formulário de Transferência Voluntária
